started refactoring the background file
