+++ src/background.ts
@@ -1,7 +1,6 @@
 import {
   DEFAULT_BORDER_SIZE,
   DEFAULT_BORDER_STYLE,
-  DEFAULT_TAB_STATE,
   ICON_PATHS,
 } from './scripts/constants';
 import {
@@ -10,525 +9,13 @@ import {
   hasPermission,
   isChromeTabClosedError,
 } from './scripts/helpers';
-import { getTimestampedScreenshotFilename } from './scripts/utils/filename';
 import Logger from './scripts/utils/logger';
-import type { TabState, TabStateChangeOptions } from './types/background';
-
-// In-memory cache for tab states. This helps reduce repeated calls to storage
-const cachedTabStates: Record<string, TabState> = {}; // tabId: TabState
-
-/**
- * Retrieves the extension state for the specified tab ID.
- * Checks cache first, then storage. Updates cache from storage.
- *
- * @param tabId - The ID of the tab to retrieve the state for.
- * @returns The state of the extension for the specified tab ID.
- */
-async function getTabState(tabId: number): Promise<TabState> {
-  const tabIdString = tabId.toString();
-
-  // Check if state exists in cache first
-  if (cachedTabStates?.[tabIdString]) {
-    return cachedTabStates[tabIdString];
-  }
-
-  try {
-    // Retrieve state from storage if not in cache
-    const storedData = await chrome.storage.local.get(tabIdString);
-    const tabState = storedData?.[tabIdString];
-
-    // Update cache with retrieved state
-    cachedTabStates[tabIdString] = tabState ?? { ...DEFAULT_TAB_STATE };
-    return cachedTabStates[tabIdString];
-  } catch (error) {
-    Logger.error(
-      `Error retrieving tab state for tab ${tabId} from storage:`,
-      error,
-    );
-    // Return default state on error
-    return { ...DEFAULT_TAB_STATE };
-  }
-}
-
-/**
- * Updates and stores the extension states for a specified tab.
- * Updates the cache and storage.
- *
- * @param options - Options to set the tab states.
- */
-async function setTabState({
-  tabId,
-  states,
-}: TabStateChangeOptions): Promise<void> {
-  const tabIdString = tabId.toString();
-
-  // Get current state (this will populate the cache if not already populated)
-  const currentState = await getTabState(tabId);
-
-  // Merge the new states into the existing cached states
-  cachedTabStates[tabIdString] = { ...currentState, ...states };
-
-  // Persist the updated state to storage
-  try {
-    await chrome.storage.local.set({
-      [tabIdString]: cachedTabStates[tabIdString],
-    });
-    Logger.info(
-      `Updated tab state for tab ${tabId} in storage:`,
-      cachedTabStates[tabIdString],
-    );
-  } catch (error) {
-    Logger.error(`Error setting tab state for tab ${tabId} in storage:`, error);
-  }
-}
-
-/**
- * Updates the extension state (icon and title) based on the active tab's state.
- * The extension is considered enabled if any mode is enabled for the active tab.
- *
- * @param tabId - The ID of the active tab.
- */
-async function updateExtensionState(tabId: number): Promise<void> {
-  try {
-    const tab = await chrome.tabs.get(tabId);
-    if (!tab?.url) return;
-
-    const isRestricted = isRestrictedUrl(tab.url);
-    const tabState = await getTabState(tabId);
-    const isActive =
-      tabState.borderMode ||
-      tabState.inspectorMode ||
-      tabState.measurementMode ||
-      tabState.rulerMode;
-
-    // Set the extension title
-    const title = isRestricted
-      ? 'Border Patrol - Restricted'
-      : isActive
-        ? 'Border Patrol - Active'
-        : 'Border Patrol - Inactive';
-
-    // Set the extension title
-    await chrome.action.setTitle({ tabId, title });
-
-    // Set the extension icon
-    const iconPath =
-      isRestricted || !isActive ? ICON_PATHS.iconDisabled : ICON_PATHS.icon16;
-
-    await chrome.action.setIcon({ tabId, path: iconPath });
-  } catch (error) {
-    if (isChromeTabClosedError(error)) {
-      Logger.warn(`Tab ${tabId} has been closed or is invalid.`);
-    } else {
-      Logger.error(`Error updating extension state for tab ${tabId}:`, error);
-    }
-
-    // Fallback to default state
-    await chrome.action.setTitle({ tabId, title: 'Border Patrol - Error' });
-    await chrome.action.setIcon({ tabId, path: ICON_PATHS.iconDisabled });
-  }
-}
-
-/**
- * Injects the content scripts and CSS into a tab if needed.
- * This should only happen once per tab session (on load/activate).
- *
- * @param tabId - The ID of the tab.
- */
-async function ensureScriptIsInjected(tabId: number): Promise<void> {
-  if (!tabId) return;
-
-  // Check if the tab is a valid webpage
-  const tab = await chrome.tabs.get(tabId);
-  if (!tab?.url || isRestrictedUrl(tab.url)) return;
-
-  try {
-    // Check if scripts are already injected by sending a ping
-    await chrome.tabs.sendMessage(tabId, { action: 'PING' });
-    // If we get a response, scripts are already injected
-    Logger.info(`Scripts already injected in tab ${tabId}.`);
-    return;
-  } catch (error) {
-    Logger.info(`Scripts likely not injected in tab ${tabId}.`);
-    // Continue with injection
-  }
-
-  try {
-    // Inject overlay styles into the active tab
-    await chrome.scripting.insertCSS({
-      target: { tabId },
-      files: ['scripts/main-content.css'],
-    });
-
-    // Inject border.js and overlay.js into the active tab
-    await chrome.scripting.executeScript({
-      target: { tabId },
-      files: ['scripts/main-content.js'],
-    });
-
-    Logger.info(`Injected content scripts and CSS into tab ${tabId}`);
-  } catch (error) {
-    if (isChromeTabClosedError(error)) return; // Tab has been closed
-    Logger.error(`Error injecting scripts or CSS into tab ${tabId}:`, error);
-  }
-}
-
-/**
- * Sends the current state of border and inspector modes, and border settings
- * to the content script in the specified tab. This should be called after state changes.
- *
- * @param tabId - The ID of the tab.
- */
-async function sendContentScriptUpdates(tabId: number): Promise<void> {
-  if (!tabId) return;
-
-  try {
-    // Retrieve the current state for the tab
-    const tabState = await getTabState(tabId);
-
-    // Send messages to update modes in the content script
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'UPDATE_BORDER_MODE',
-      isEnabled: tabState.borderMode,
-    });
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'UPDATE_INSPECTOR_MODE',
-      isEnabled: tabState.inspectorMode,
-    });
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'UPDATE_MEASUREMENT_MODE',
-      isEnabled: tabState.measurementMode,
-    });
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'UPDATE_RULER_MODE',
-      isEnabled: tabState.rulerMode,
-    });
-    Logger.info(`Sent mode updates to tab ${tabId}:`, tabState);
-
-    // Send border settings to the content script
-    const settings = await chrome.storage.local.get([
-      'borderSize',
-      'borderStyle',
-    ]);
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'UPDATE_BORDER_SETTINGS',
-      borderSize: settings.borderSize ?? DEFAULT_BORDER_SIZE,
-      borderStyle: settings.borderStyle ?? DEFAULT_BORDER_STYLE,
-    });
-    Logger.info(`Sent border settings update to tab ${tabId}:`, settings);
-  } catch (error) {
-    Logger.warn(`Error sending content script updates to tab ${tabId}:`, error);
-  }
-}
-
-/**
- * Handles state changes for a specific tab.
- * Updates storage, extension state (icon/title), and sends updates to content scripts.
- *
- * @param options - Options for handling the tab state change.
- */
-async function handleTabStateChange({
-  tabId,
-  states,
-}: TabStateChangeOptions): Promise<void> {
-  if (!tabId) return;
-
-  // Ensure scripts are injected first if needed (Chrome handles not injecting duplicates)
-  await ensureScriptIsInjected(tabId);
-
-  if (states) {
-    // Update the state in storage and cache
-    await setTabState({ tabId, states });
-  }
-
-  // Update the extension icon and title based on the NEW state
-  await updateExtensionState(tabId);
-
-  // Send the NEW state to the content script(s)
-  await sendContentScriptUpdates(tabId);
-}
-
-/**
- * Captures a visible tab and downloads the screenshot.
- *
- * @param windowId - The ID of the window containing the tab to capture.
- */
-async function captureAndDownloadScreenshot(windowId: number): Promise<void> {
-  const format = 'png';
-  try {
-    // Capture the visible tab
-    const screenshotUrl = await chrome.tabs.captureVisibleTab(windowId, {
-      format,
-    });
-    Logger.info('Screenshot captured successfully:', {
-      windowId,
-      screenshotUrl,
-    });
-
-    // Generate a filename with timestamp
-    const filename = getTimestampedScreenshotFilename(format);
-    Logger.info('Generated filename:', filename);
-
-    // Download the screenshot using the downloads API
-    await chrome.downloads.download({
-      url: screenshotUrl,
-      filename,
-      saveAs: true, // Prompt user to choose location
-      conflictAction: 'uniquify', // Add a number if filename exists
-    });
-  } catch (error) {
-    Logger.error('Error in captureAndDownloadScreenshot:', error);
-    throw error; // Re-throw to be handled by the caller
-  }
-}
-
-/**
- * Ensures that the offscreen document used for canvas stitching exists.
- * Creates it only if no offscreen context is already open.
- */
-async function ensureOffscreenDocument(): Promise<void> {
-  const url = chrome.runtime.getURL('offscreen/offscreen.html');
-  const existingContexts = await chrome.runtime.getContexts({
-    contextTypes: [chrome.runtime.ContextType.OFFSCREEN_DOCUMENT],
-    documentUrls: [url],
-  });
-  if (existingContexts.length > 0) return;
-
-  await chrome.offscreen.createDocument({
-    url,
-    reasons: [chrome.offscreen.Reason.DOM_SCRAPING],
-    justification: 'Canvas stitching for full-page screenshot',
-  });
-}
-
-/**
- * Polls the offscreen document with an OFFSCREEN_PING until it responds or
- * the maximum number of attempts is exhausted. This guards against the race
- * where createDocument() resolves before the offscreen page has registered
- * its onMessage listener.
- *
- * @param maxAttempts - Maximum ping attempts before giving up.
- * @param intervalMs - Delay between attempts in milliseconds.
- */
-async function waitForOffscreenReady(
-  maxAttempts = 20,
-  intervalMs = 50,
-): Promise<void> {
-  for (let attempt = 0; attempt < maxAttempts; attempt++) {
-    try {
-      const response = await chrome.runtime.sendMessage({
-        action: 'OFFSCREEN_PING',
-      });
-      if (response?.ready) return;
-    } catch {
-      // Offscreen listener not yet registered; retry after a short delay.
-    }
-    await sleep(intervalMs);
-  }
-  Logger.warn('Offscreen document did not become ready in time');
-}
-
-interface StitchFrame {
-  dataUrl: string;
-  x: number;
-  y: number;
-}
-
-/**
- * Returns a promise that resolves after the specified number of milliseconds.
- *
- * @param ms - Milliseconds to wait.
- */
-function sleep(ms: number): Promise<void> {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-/**
- * Scrolls a tab through its full height and width, captures each viewport,
- * stitches the frames in an offscreen document, then downloads the result.
- *
- * @param tabId - The ID of the tab to capture.
- * @param windowId - The window ID used by captureVisibleTab.
- */
-async function captureAndDownloadFullPageScreenshot(
-  tabId: number,
-  windowId: number,
-): Promise<void> {
-  const format = 'png';
-
-  // Chrome's captureVisibleTab is rate-limited to ~2 calls/second.
-  // We enforce a minimum interval between captures to avoid the quota error.
-  const MIN_CAPTURE_INTERVAL_MS = 600;
-
-  /**
-   * Sends a screenshot overlay command to the content script without breaking
-   * the capture flow if the page cannot process the message.
-   *
-   * @param action - The overlay action to send.
-   */
-  const sendScreenshotOverlayCommand = async (
-    action:
-      | 'SHOW_SCREENSHOT_OVERLAY'
-      | 'HIDE_SCREENSHOT_OVERLAY'
-      | 'REMOVE_SCREENSHOT_OVERLAY',
-  ): Promise<void> => {
-    try {
-      await chrome.tabs.sendMessage(tabId, { action });
-    } catch (error) {
-      Logger.warn(`Full-page capture overlay command failed: ${action}`, error);
-    }
-  };
-
-  /**
-   * Best-effort fallback that removes the overlay host directly from the page
-   * if content-script messaging cannot complete cleanup.
-   */
-  const forceRemoveScreenshotOverlay = async (): Promise<void> => {
-    try {
-      await chrome.scripting.executeScript({
-        target: { tabId },
-        func: () => {
-          const container = document.getElementById(
-            'bp-screenshot-overlay-container',
-          );
-          container?.remove();
-        },
-      });
-    } catch (error) {
-      Logger.warn('Full-page capture overlay fallback removal failed', error);
-    }
-  };
-
-  // Ensure the fullpage content script is present before sending any messages.
-  // On freshly-installed or never-activated tabs, the listener won't exist yet.
-  await ensureScriptIsInjected(tabId);
-
-  // 1. Get full page dimensions and save the original scroll position.
-  const dims = (await chrome.tabs.sendMessage(tabId, {
-    action: 'GET_PAGE_DIMENSIONS',
-  })) as {
-    scrollHeight: number;
-    scrollWidth: number;
-    viewportHeight: number;
-    viewportWidth: number;
-    scrollX: number;
-    scrollY: number;
-    devicePixelRatio: number;
-  };
-
-  Logger.info('Full-page capture: page dimensions', dims);
-
-  const {
-    scrollHeight,
-    scrollWidth,
-    viewportHeight,
-    viewportWidth,
-    scrollX: origX,
-    scrollY: origY,
-    devicePixelRatio,
-  } = dims;
-
-  const frames: StitchFrame[] = [];
-
-  // Hide fixed/sticky elements so they don't repeat in every captured frame.
-  await chrome.tabs.sendMessage(tabId, { action: 'HIDE_FIXED_ELEMENTS' });
-
-  // Show the progress overlay so the user knows capture is underway and
-  // accidental clicks cannot interrupt the process.
-  await sendScreenshotOverlayCommand('SHOW_SCREENSHOT_OVERLAY');
-
-  try {
-    // 2. Loop through rows and columns, capturing each viewport-sized region.
-    let lastCaptureTime = 0;
-
-    for (let y = 0; y < scrollHeight; y += viewportHeight) {
-      for (let x = 0; x < scrollWidth; x += viewportWidth) {
-        // Scroll the page and wait for repaint.
-        const actual = (await chrome.tabs.sendMessage(tabId, {
-          action: 'SCROLL_TO',
-          x,
-          y,
-        })) as { scrollX: number; scrollY: number };
-
-        // Enforce the minimum interval between captureVisibleTab calls.
-        const elapsed = Date.now() - lastCaptureTime;
-        if (elapsed < MIN_CAPTURE_INTERVAL_MS) {
-          await sleep(MIN_CAPTURE_INTERVAL_MS - elapsed);
-        }
-
-        // Keep the overlay visible while the page scrolls and during the quota
-        // wait, then hide it immediately before capture so it does not appear
-        // in the saved image.
-        await sendScreenshotOverlayCommand('HIDE_SCREENSHOT_OVERLAY');
-
-        // Capture the visible area after scrolling.
-        const dataUrl = await chrome.tabs.captureVisibleTab(windowId, {
-          format,
-        });
-        lastCaptureTime = Date.now();
-
-        frames.push({ dataUrl, x: actual.scrollX, y: actual.scrollY });
-        Logger.info(
-          `Full-page capture: frame at (${actual.scrollX}, ${actual.scrollY})`,
-        );
-
-        // Restore the overlay so the user can see the page scroll to the next
-        // position before the next capture starts.
-        await sendScreenshotOverlayCommand('SHOW_SCREENSHOT_OVERLAY');
-      }
-    }
-  } finally {
-    // 3. Remove the overlay and restore fixed/sticky elements and the
-    //    original scroll position.
-    await sendScreenshotOverlayCommand('REMOVE_SCREENSHOT_OVERLAY');
-    await forceRemoveScreenshotOverlay();
-    await chrome.tabs.sendMessage(tabId, { action: 'RESTORE_FIXED_ELEMENTS' });
-    await chrome.tabs.sendMessage(tabId, {
-      action: 'RESTORE_SCROLL',
-      x: origX,
-      y: origY,
-    });
-  }
-
-  // 4. Stitch frames in the offscreen document.
-  await ensureOffscreenDocument();
-  await waitForOffscreenReady();
-
-  const stitchResponse = await new Promise<{
-    dataUrl?: string;
-    error?: string;
-  }>(resolve => {
-    chrome.runtime.sendMessage(
-      {
-        action: 'STITCH_FRAMES',
-        frames,
-        totalWidth: scrollWidth,
-        totalHeight: scrollHeight,
-        viewportWidth,
-        viewportHeight,
-        devicePixelRatio,
-      },
-      response => resolve(response),
-    );
-  });
-
-  if (stitchResponse.error || !stitchResponse.dataUrl) {
-    throw new Error(
-      `Full-page stitch failed: ${stitchResponse.error ?? 'no data URL returned'}`,
-    );
-  }
-
-  // 5. Download the stitched image.
-  const filename = getTimestampedScreenshotFilename(format);
-  await chrome.downloads.download({
-    url: stitchResponse.dataUrl,
-    filename,
-    saveAs: true,
-    conflictAction: 'uniquify',
-  });
-
-  Logger.info('Full-page screenshot downloaded:', filename);
-}
+import { handleTabStateChange } from './background/extension-ui';
+import {
+  captureAndDownloadScreenshot,
+  captureAndDownloadFullPageScreenshot,
+} from './background/screenshot';
+import { clearTabStateCache, getTabState } from './background/tab-state';
 
 /**
  * Creates the "Border Patrol" context menu with sub-items to toggle each mode.
@@ -688,7 +175,7 @@ chrome.tabs.onRemoved.addListener(
     Logger.info('onRemoved', tabId, removeInfo);
 
     // Remove the tab state from the cache
-    delete cachedTabStates[tabId];
+    clearTabStateCache(tabId);
 
     // Optionally clear storage for the tab if needed
     try {
