+++ src/background/screenshot/index.ts
@@ -1,0 +1,2 @@
+export { captureAndDownloadScreenshot } from './visible';
+export { captureAndDownloadFullPageScreenshot } from './fullpage';
